requirements.txt
fastapi
uvicorn
pyswisseph
geopy